add basic auth on hxr adapter
