+++ lib/adapters/http.js
@@ -3,6 +3,7 @@
 var utils = require('./../utils');
 var settle = require('./../core/settle');
 var buildURL = require('./../helpers/buildURL');
+var parseUrl = require('./../helpers/parseUrl');
 var http = require('http');
 var https = require('https');
 var httpFollow = require('follow-redirects').http;
@@ -45,24 +46,9 @@ module.exports = function httpAdapter(config) {
       headers['Content-Length'] = data.length;
     }
 
-    // HTTP basic authentication
-    var auth = undefined;
-    if (config.auth) {
-      var username = config.auth.username || '';
-      var password = config.auth.password || '';
-      auth = username + ':' + password;
-    }
-
-    // Parse url
-    var parsed = url.parse(config.url);
-    var protocol = parsed.protocol || 'http:';
-
-    if (!auth && parsed.auth) {
-      var urlAuth = parsed.auth.split(':');
-      var urlUsername = urlAuth[0] || '';
-      var urlPassword = urlAuth[1] || '';
-      auth = urlUsername + ':' + urlPassword;
-    }
+    var parsed = parseUrl(config);
+    var auth = parsed.auth;
+    var protocol = parsed.protocol;
 
     if (auth) {
       delete headers.Authorization;
