+++ lib/adapters/xhr.js
@@ -2,6 +2,7 @@
 
 var utils = require('./../utils');
 var settle = require('./../core/settle');
+var parseUrl = require('./../helpers/parseUrl');
 var buildURL = require('./../helpers/buildURL');
 var parseHeaders = require('./../helpers/parseHeaders');
 var isURLSameOrigin = require('./../helpers/isURLSameOrigin');
@@ -35,11 +36,11 @@ module.exports = function xhrAdapter(config) {
       request.ontimeout = function handleTimeout() {};
     }
 
+    var auth = parseUrl(config).auth;
+
     // HTTP basic authentication
-    if (config.auth) {
-      var username = config.auth.username || '';
-      var password = config.auth.password || '';
-      requestHeaders.Authorization = 'Basic ' + btoa(username + ':' + password);
+    if (auth) {
+      requestHeaders.Authorization = 'Basic ' + btoa(auth);
     }
 
     request.open(config.method.toUpperCase(), buildURL(config.url, config.params, config.paramsSerializer), true);
